Consolidating compressible gas examples
advect1d, advect1d-newapi, sod-shock, and rayleigh-taylor
all used variations on the same gas_compress.cxx model.
These are all now under examples/gas-compress and use the same
code with different inputs.
