--- examples/advect1d/data/BOUT.inp
@@ -1,99 +1,0 @@
-# settings file for BOUT++
-#
-# Simple 1D advection test case
-#
-# By default everything in NetCDF.
-# To change to PDB format, change the settings to
-#   grid="data/advect.grd.pdb"
-#   dump_format = "pdb"
-
-##################################################
-# settings used by the core code
-
-NOUT = 10       # number of time-steps
-TIMESTEP = 1.0e-2   # time between outputs
-
-ShiftXderivs = false # use shifted radial derivatives?
-TwistShift = false  # use twist-shift condition?
-
-MZ = 3     # number of points in z direction (2^n + 1)
-ZMIN = 0.0
-ZMAX = 0.25
-
-MXG = 2
-MYG = 2
-
-grid="data/advect.grd.cdl"  # Grid file
-
-dump_format = "nc" # Output format. nc = NetCDF
-
-##################################################
-# derivative methods
-
-[mesh:ddx]
-
-first = C4  # order of first x derivatives (options are 2 or 4)
-second = C4 # order of second x derivatives (2 or 4)
-upwind = U1 # order of upwinding method (1, 4, or 0 = TVD)
-
-[mesh:ddy]
-
-first = C4
-second = C2
-upwind = U1
-
-[mesh:ddz]
-
-first = C4
-second = C2
-upwind = U1
-
-##################################################
-# Solver settings
-
-[solver]
-
-# mudq, mldq, mukeep, mlkeep preconditioner options
-ATOL = 1.0e-10 # absolute tolerance
-RTOL = 1.0e-5  # relative tolerance
-
-##################################################
-# settings for gas_compress
-
-[gas]
-
-include_viscosity = false
-viscosity = 0.0
-
-v0_multiply = -1.0
-
-[All]
-scale = 0.0 # default size of initial perturbations
-
-# form of initial profile:
-# 0 - constant
-# 1 - Gaussian
-# 2 - Sinusoidal
-# 3 - Mix of mode numbers (like original BOUT)
-
-xs_opt = 0
-
-ys_opt = 1
-zs_opt = 2
-
-ys_s0 = 0.5
-ys_wd = 0.1
-
-zs_mode = 1 # toroidal mode number
-
-# Boundary conditions
-# dirichlet = zero value
-# neumann = zero gradient
-
-bndry_all = neumann # Set all boundaries to Neumann
-
-[density]
-scale = 0.5
-
-[solver]
-type=petsc-3.2
