+++ examples/gas-compress/sod-shock/BOUT.inp
@@ -1,37 +1,22 @@
 # settings file for BOUT++
 #
-# Simple 1D advection test case
+# Sod shock problem in 1D
 #
 
 ##################################################
 # settings used by the core code
 
-NOUT = 10       # number of time-steps
+NOUT = 20       # number of time-steps
 TIMESTEP = 1.0e-2   # time between outputs
 
-ShiftXderivs = false # use shifted radial derivatives?
-TwistShift = false  # use twist-shift condition?
-
-MZ = 3     # number of points in z direction (2^n + 1)
-ZMIN = 0.0
-ZMAX = 0.25
+MZ = 1
 
 MXG = 2
 MYG = 2
 
-dump_format = "nc" # Output format. nc = NetCDF
+grid="sod.grd.nc"  # Grid file
 
-[mesh]
-
-ixseps1 = 5
-ixseps2 = 5
-nx = 5
-ny = 128
-dx = 1.
-dy = 0.0078125
-vy = 1.
-pressure = 0.6
-density = 1.0
+StaggerGrids = false
 
 ##################################################
 # derivative methods
@@ -44,9 +29,10 @@ upwind = U1 # order of upwinding method (1, 4, or 0 = TVD)
 
 [mesh:ddy]
 
-first = C4
+first = C2
 second = C2
 upwind = U1
+flux = split
 
 [mesh:ddz]
 
@@ -97,7 +83,6 @@ zs_mode = 1 # toroidal mode number
 # neumann = zero gradient
 
 bndry_all = neumann # Set all boundaries to Neumann
+bndry_target = neumann_2ndorder
 
 [density]
-scale = 0.5
-
