--- examples/rayleigh-taylor/makefile
@@ -1,6 +1,0 @@
-
-BOUT_TOP	= ../..
-
-SOURCEC		= gas_compress.cxx
-
-include $(BOUT_TOP)/make.config
